requirements.txt
astroid==2.3.3
autopep8==1.4.4
esptool==3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.3
pymongo==3.9.0
pyserial==3.4
python-dateutil==2.8.1
six==1.13.0
typed-ast==1.4.3
websockets==8.1
wrapt==1.11.2
